// Frontend/generated/models/agendaTask.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Backend | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AgendaItem } from './agendaItem';
import type { Department } from './department';
import type { User } from './user';

export interface AgendaTask {
  id?: string;
  name: string;
  /** @nullable */
  description?: string | null;
  deadlineDate: string;
  agendaItemId: string;
  agendaItem?: null | AgendaItem;
  /** @nullable */
  departmentId?: string | null;
  department?: null | Department;
  /** @nullable */
  userId?: string | null;
  user?: null | User;
}